Fix single-precision truncation in HalfType that broke sqrt() for integer operands
HalfType::eval() (halftype.h) returned 0.5f — a float literal. This
caused Power::eval (power.h:48-54) to deduce its working type NT as:

    NT = decltype(NT1{} * float{})

When NT1 was double (field expressions), this promoted to double
harmlessly. But when NT1 was int64_t (e.g. WaveNumber::eval returns
IdxArray<NDim> = array<int64_t, NDim>), the promotion went:

    int64_t * float → float

forcing pow(float, float) — a 32-bit sqrt. The result, cast back to
double upon field assignment, carried ~2.4e-08 of truncation error.
This exceeded AlmostEqual's epsilon of ~1.49e-08, causing
test-wavenumber to fail on the norm check for all NDim values while
every other check (components, norm2, imaginary parts) passed.

Fix: change the literal from 0.5f to 0.5 so HalfType::eval returns
double. Now int64_t * double → double, and pow() executes in 64-bit
precision. Float-precision builds (HAVE_FFTFLOAT) are unaffected:
the double result is narrowed to float at field assignment, matching
the previous behavior.

+++ include/TempLat/lattice/algebra/constants/halftype.h
@@ -22,11 +22,11 @@ namespace TempLat
       requires IsVariadicIndex<IDX...>
     DEVICE_INLINE_FUNCTION static constexpr auto eval(const IDX &...i)
     {
-      return 0.5f;
+      return 0.5;
     }
     inline explicit operator double() const { return value; }
     inline explicit operator float() const { return value; }
-    static constexpr float value = 0.5f;
+    static constexpr double value = 0.5;
   };
 } // namespace TempLat
